+++ src/OF_WAR.cpp
@@ -290,11 +290,12 @@ void FirmWar::disp_build_menu(int refreshFlag)
 		if( !added_count || queue_weapon_selected >= added_count)
 			queue_weapon_selected = -1;
 
-		button_cancel.paint(x, y, x+BUILD_BUTTON_WIDTH-1, y+BUILD_BUTTON_HEIGHT*3/4,
+		button_cancel.paint(x, y, x+Ambition::Draw::calculateDoneButtonWidth(BUILD_BUTTON_WIDTH, this)-1, y+BUILD_BUTTON_HEIGHT*3/4,
 		ButtonCustom::disp_text_button_func, ButtonCustomPara((void*)_("Done"),0) );
 	}
 	// ###### end Gilbert 10/9 ########//
 
+	Ambition::Building::printProductionQueueTotals(this, refreshFlag, button_cancel.x2, button_cancel.y1, button_cancel.y2);
 }
 //----------- End of function FirmWar::disp_build_menu -----------//